Zone is working. A zone is a collection of cells in which a cpu has access to. It also list the entities found. CPUs can and should share the same zone (Arc type) if they can find eachother

+++ src/game/game_settings.rs
@@ -3,7 +3,7 @@ use std::error::Error;
 use crate::{game::map::map_settings::MapSettings, settings::settings::Settings};
 
 pub struct GameSettings {
-    pub nb_humans: u32,
+    pub nb_humans: usize,
     pub map_settings: MapSettings,
     // TODO: complete with more options that the user can chose before the game:
     // powers up list?
@@ -22,7 +22,7 @@ impl GameSettings {
     pub fn new(settings: &Settings, map_settings: MapSettings) -> Result<Self, Box<dyn Error>> {
         Ok(Self {
             map_settings,
-            nb_humans: settings.binds.len() as u32,
+            nb_humans: settings.binds.len(),
         })
     }
 
@@ -30,7 +30,7 @@ impl GameSettings {
         self.map_settings = map_settings;
     }
 
-    pub fn set_nb_humans(&mut self, nb_humans: u32) {
+    pub fn set_nb_humans(&mut self, nb_humans: usize) {
         self.nb_humans = nb_humans;
     }
 }
